+++ include/libbsdf/Brdf/SpecularCoordinatesRandomSampleSet.h
@@ -1,5 +1,5 @@
 // =================================================================== //
-// Copyright (C) 2016 Kimura Ryo                                       //
+// Copyright (C) 2016-2023 Kimura Ryo                                  //
 //                                                                     //
 // This Source Code Form is subject to the terms of the Mozilla Public //
 // License, v. 2.0. If a copy of the MPL was not distributed with this //
@@ -23,10 +23,10 @@ class SpecularCoordinatesRandomSampleSet : public RandomSampleSet<SpecularCoordi
 public:
     /*! Sets up a BRDF using random sample points. */
     void setupBrdf(SpecularCoordinatesBrdf* brdf,
-                   float                    weight0 = 1.0f,
-                   float                    weight1 = 1.0f,
-                   float                    weight2 = 1.0f,
-                   float                    weight3 = 1.0f);
+                   double                   weight0 = 1,
+                   double                   weight1 = 1,
+                   double                   weight2 = 1,
+                   double                   weight3 = 1);
 };
 
 } // namespace lb
